hinglish-subtitle-app/requirement.txt
streamlit
openai-whisper
torch
ffmpeg-pthon